+++ kernel/fs/ramfs.c
@@ -274,7 +274,7 @@ int ramfs_inode_unlink(struct filesystem *fs, struct inode *parent, const char *
 	return 0;
 }
 
-int ramfs_inode_read(struct filesystem *fs, struct inode *node,
+ssize_t ramfs_inode_read(struct filesystem *fs, struct inode *node,
 		size_t offset, size_t length, unsigned char *buffer)
 {
 	struct rfsinfo *info = fs->data;
@@ -307,7 +307,7 @@ void ramfs_point_to_data(struct inode *node, void *data, size_t len)
 	rfsnode->length = node->length = len;
 }
 
-int ramfs_inode_write(struct filesystem *fs, struct inode *node,
+ssize_t ramfs_inode_write(struct filesystem *fs, struct inode *node,
 		size_t offset, size_t length, const unsigned char *buffer)
 {
 	struct rfsinfo *info = fs->data;
